Added duplicate value handler
Added functionality to prevent the user from adding a value to the list if that value is already in the list. Simply preventing this makes implementation less complicated. 

In the future I may add a "duplicate" variable to each node, which will act as a flag and allow the user to add duplicate values while retaining a list that is logically sorted.